LayoutTests/security/clip-invalid-rect-2-expected.txt

ALERT: Test passed because nothing crashed
some text
